Update for SLES
RedHat has decided to kill CentOS, so in order to support the future we
should update the tasks to ensure they target a more stable
distribution. SLES is a nice, stable, widely-available, RPM-based
distro, and is available on most clouds for no fee. Seems like an
excellent CentOS replacement for me.

I've tried to genericise as much as possible, in any case, so as long as
you use firewalld it should work on basically anything.